Check remote API methods arguments count
This commit resolves the issue when a client could break the server by
calling a method with incorrect number of arguments causing a TypeError
exception ("callback is not a function").

Later a full-blown IDL for remote APIs (#608) will be implemented in
JSTP, so functions will be able to have optional parameters and
parameters with default values and JSTP will check the types of
arguments.