Use symlinks to JSTP browser libs
Instead of copying browser JSTP libraries from the JSTP distribution on
`npm install`, use symlinks to them so that the client-side part of
`metarhia-jstp` is always up-to-date when running an example
application and always has the same version as Impress uses at
server side.

This kind of refs #697. Though the described bug is caused by another
issue, a user could theoretically run into similar behavior because of
this issue too (though the symptoms would be rather opposite: new
interface in the example application and old interface in the included
JavaScript library).

Refs: https://github.com/metarhia/Impress/issues/697