Remove warning on safe require
We can't print warning before logger is loaded. In future we will refactor this
but now it crashes server so this safe warning is not critical and can be
removed. Detecting not loaded modules is easy because accessing them we will
get additional error and it will be recorded to logs.