Fix parent test timeout in test mode
As of now metatests uses 30s as default test timeout therefore, because
impress in core.js when starting the subtests didn't change the timeout
value all tests were limited to being run no more that 30 seconds
in total and after this limit was reached would be forcefully finished.

This is obviously not enough for an extensive test suite.
This commit allows to specify a custom testTimeout in the application
config and disables timeout on the Impress parent test.