Allow to opt-out of impress-cli installation
Current impress-cli installation command requires root access to be
executed which is displayed during the installation process as a sudo
prompt without any explanation whick is confusing and would most likely
result in halting the command execution.
This commit aks the user if the cli needs to be installed first.

This also removes the installation of impress-cli on non-TTY terminals
as it may be run under a non-root user which will result in sudo prompt
with nobody to write the password and hang whole installation.